[Issue #2] Format functions should be defined/referenced from the class so they can be overridden
