+++ CHANGELOG.md
@@ -3,7 +3,6 @@
 ## [1.1.0]
 
 ### Additions
-
-### Fixes
+- [Issue #2] Format functions should be defined/referenced from the class so they can be overridden
 
 ## [1.0.0] Initial Release
