--- src/format-entry/formatted-entry-interface.ts
@@ -1,7 +1,0 @@
-/**
- * Holds winston-friendly arguments from formatEntry()
- */
-export interface FormattedEntryInterface {
-	message?: string;
-	data?: any | any[];
-}
