--- src/format-entry/index.ts
@@ -1,2 +1,0 @@
-export { formatEntry } from './format-entry';
-export { FormattedEntryInterface } from './formatted-entry-interface';
